move utility to datalayer folder
